3-B/Database-System-Principles-Curriculum-Design/database-backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Mail
psycopg2-binary
python-dotenv
flasgger
pandas
openpyxl
pytest